remove comments questioning whether we could use x.cv and x.cc in the conditions (we can't)
Consider if (x.cc <= zero) { midcv = x.cc; midcc = x.cv; }

Then for x_mc = [-2, (-1, -0.5), 3] we would enter this first branch but midcc is incorrect (should also be x.cc). This is because we then assume that the line segment for the overapproximation could just
go from -1 to -0.5 (hence going downwards) even though it must go from -2 to 3 (hence having a positive steepness).